the beginnings of the map editor interface
This code commit lays down the structure for the Backbone/Polymaps/d3
application.

At the moment, GeoPoint's are drawn on the map. The code for drawing
Segment's on the map is in progress. (I can't figure out how to
properly transform the coordinates for display on the Polymaps map.)

Also I've added a basic API interface to produce JSON for consumption
by Backbone. Note that data like GeoPoint's and Segment's are not being
accessed with respect to a parent PedProject just yet. And the API
doesn't perform any sort of authentication or authorization.