adding users/projects/organizations and authentication
All this code comes from the OrgUP app. Originally, I had planned that
to be a separate app that could serve as a backend service to PedPlus
and to VizPlus. But it's turning out to be more complicated than I
expected to tie PedPlus to OrgUP via OAuth, so instead I'm just making
this one Web app for now. In the future, when we do VizPlus, we can
break them apart again.