add an internal method for changing the radius of GeoPoint's and the stroke-width of Segment's
This is so that Victor can have a way to change the radius of
GeoPoint's in order to do detailed editing for tight quarters (by
opening the Chrome developer tools panel). In the future, this is
functionality that's worth exposing to end users through a project
settings window.

Note that in the process I found that I wasn't properly applying the
transform function in the SegmentLayer.load and GeoPoint.load
functions. I've now added that, following this example sent to me by
Mike Bostock: http://bl.ocks.org/645734